Module aws_sdk_lookoutmetrics::error

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

too_many_requests_exception

See TooManyRequestsException

validation_exception

See ValidationException

Structs

AccessDeniedException

You do not have sufficient permissions to perform this action.

ActivateAnomalyDetectorError

BackTestAnomalyDetectorError

ConflictException

There was a conflict processing the request. Try your request again.

CreateAlertError

CreateAnomalyDetectorError

CreateMetricSetError

DeleteAlertError

DeleteAnomalyDetectorError

DescribeAlertError

DescribeAnomalyDetectionExecutionsError

DescribeAnomalyDetectorError

DescribeMetricSetError

GetAnomalyGroupError

GetFeedbackError

GetSampleDataError

InternalServerException

The request processing has failed because of an unknown error, exception, or failure.

ListAlertsError

ListAnomalyDetectorsError

ListAnomalyGroupSummariesError

ListAnomalyGroupTimeSeriesError

ListMetricSetsError

ListTagsForResourceError

PutFeedbackError

ResourceNotFoundException

The specified resource cannot be found. Check the ARN of the resource and try again.

ServiceQuotaExceededException

The request exceeded the service's quotas. Check the service quotas and try again.

TagResourceError

TooManyRequestsException

The request was denied due to too many requests being submitted at the same time.

UntagResourceError

UpdateAnomalyDetectorError

UpdateMetricSetError

ValidationException

The input fails to satisfy the constraints specified by the AWS service. Check your input values and try

Enums

ActivateAnomalyDetectorErrorKind

BackTestAnomalyDetectorErrorKind

CreateAlertErrorKind

CreateAnomalyDetectorErrorKind

CreateMetricSetErrorKind

DeleteAlertErrorKind

DeleteAnomalyDetectorErrorKind

DescribeAlertErrorKind

DescribeAnomalyDetectionExecutionsErrorKind

DescribeAnomalyDetectorErrorKind

DescribeMetricSetErrorKind

GetAnomalyGroupErrorKind

GetFeedbackErrorKind

GetSampleDataErrorKind

ListAlertsErrorKind

ListAnomalyDetectorsErrorKind

ListAnomalyGroupSummariesErrorKind

ListAnomalyGroupTimeSeriesErrorKind

ListMetricSetsErrorKind

ListTagsForResourceErrorKind

PutFeedbackErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateAnomalyDetectorErrorKind

UpdateMetricSetErrorKind