Module aws_sdk_lookoutmetrics::model

Modules

action

See Action

alert

See Alert

alert_summary

See AlertSummary

anomaly_detector_config

See AnomalyDetectorConfig

anomaly_detector_config_summary

See AnomalyDetectorConfigSummary

anomaly_detector_summary

See AnomalyDetectorSummary

anomaly_group

See AnomalyGroup

anomaly_group_statistics

See AnomalyGroupStatistics

anomaly_group_summary

See AnomalyGroupSummary

anomaly_group_time_series

See AnomalyGroupTimeSeries

anomaly_group_time_series_feedback

See AnomalyGroupTimeSeriesFeedback

app_flow_config

See AppFlowConfig

cloud_watch_config

See CloudWatchConfig

contribution_matrix

See ContributionMatrix

csv_format_descriptor

See CsvFormatDescriptor

dimension_contribution

See DimensionContribution

dimension_name_value

See DimensionNameValue

dimension_value_contribution

See DimensionValueContribution

execution_status

See ExecutionStatus

file_format_descriptor

See FileFormatDescriptor

itemized_metric_stats

See ItemizedMetricStats

json_format_descriptor

See JsonFormatDescriptor

lambda_configuration

See LambdaConfiguration

metric

See Metric

metric_level_impact

See MetricLevelImpact

metric_set_summary

See MetricSetSummary

metric_source

See MetricSource

rds_source_config

See RdsSourceConfig

redshift_source_config

See RedshiftSourceConfig

s3_source_config

See S3SourceConfig

sample_data_s3_source_config

See SampleDataS3SourceConfig

sns_configuration

See SnsConfiguration

time_series

See TimeSeries

time_series_feedback

See TimeSeriesFeedback

timestamp_column

See TimestampColumn

validation_exception_field

See ValidationExceptionField

vpc_configuration

See VpcConfiguration

Structs

Action

A configuration that specifies the action to perform when anomalies are detected.

Alert

A configuration for Amazon SNS-integrated notifications.

AlertSummary

Provides a summary of an alert's configuration.

AnomalyDetectorConfig

Contains information about a detector's configuration.

AnomalyDetectorConfigSummary

Contains information about a detector's configuration.

AnomalyDetectorSummary

Contains information about an an anomaly detector.

AnomalyGroup

A group of anomalous metrics

AnomalyGroupStatistics

Aggregated statistics for a group of anomalous metrics.

AnomalyGroupSummary

Details about a group of anomalous metrics.

AnomalyGroupTimeSeries

An anomalous metric in an anomaly group.

AnomalyGroupTimeSeriesFeedback

Feedback for an anomalous metric.

AppFlowConfig

Details about an Amazon AppFlow flow datasource.

CloudWatchConfig

Details about an Amazon CloudWatch datasource.

ContributionMatrix

Details about dimensions that contributed to an anomaly.

CsvFormatDescriptor

Contains information about how a source CSV data file should be analyzed.

DimensionContribution

Details about a dimension that contributed to an anomaly.

DimensionNameValue

A dimension name and value.

DimensionValueContribution

The severity of a value of a dimension that contributed to an anomaly.

ExecutionStatus

The status of an anomaly detector run.

FileFormatDescriptor

Contains information about a source file's formatting.

ItemizedMetricStats

Aggregated statistics about a measure affected by an anomaly.

JsonFormatDescriptor

Contains information about how a source JSON data file should be analyzed.

LambdaConfiguration

Contains information about a Lambda configuration.

Metric

A calculation made by contrasting a measure and a dimension from your source data.

MetricLevelImpact

Details about a measure affected by an anomaly.

MetricSetSummary

Contains information about a dataset.

MetricSource

Contains information about source data used to generate a metric.

RdsSourceConfig

Contains information about the Amazon Relational Database Service (RDS) configuration.

RedshiftSourceConfig

Provides information about the Amazon Redshift database configuration.

S3SourceConfig

Contains information about the configuration of the S3 bucket that contains source files.

SampleDataS3SourceConfig

Contains information about the source configuration in Amazon S3.

SnsConfiguration

Contains information about the SNS topic to which you want to send your alerts and the IAM role that has

TimeSeries

Details about a metric. A metric is an aggregation of the values of a measure for a dimension value, such as

TimeSeriesFeedback

Details about feedback submitted for an anomalous metric.

TimestampColumn

Contains information about the column used to track time in a source data file.

ValidationExceptionField

Contains information about a a field in a validation exception.

VpcConfiguration

Contains configuration information about the Amazon Virtual Private Cloud (VPC).

Enums

AggregationFunction

AlertStatus

AlertType

AnomalyDetectionTaskStatus

AnomalyDetectorStatus

CsvFileCompression

Frequency

JsonFileCompression

ValidationExceptionReason